Run tests in Node.js 4, 6, 8 and 9 by switching to CircleCI workflows.
This aims to accomplish the request @glasser made in #161, which I came
across while reviewing docs PRs.

My recommendation is that the original intention of @evans PR should be broken
out into its own commit/PR which only changes the `target` transpilation
language to `es5`, after (and of course, "if") this PR has landed.

I expect the results of this commit's first test-run on CircleCI to support
the original need for #161 by showing a failure on Node.js 4, though I
wouldn't be entirely surprised to see failures on  Node.js 6 either.

Ref: https://github.com/apollographql/apollo-engine-js/pull/161#pullrequestreview-105205818